requirements.txt
pika==1.3.0
opencv-python==4.6.0.66
python-dotenv==0.21.0
PyMySQL==1.0.2